Add arguments for fetching Alerts, break off getFeedMessage to another service
