+++ src/realtime/realtime.service.ts
@@ -1,39 +1,16 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import fetch from 'node-fetch';
 import { FeedEntity, FeedMessage } from 'proto/gtfs-realtime';
 import { getAlertUrls, getConfigByFeedIndex } from 'util/';
 import { AlertEntity } from 'realtime/entities/alert.entity';
+import { FeedService } from 'realtime/feed.service';
 
 @Injectable()
 export class RealtimeService {
-  constructor(private readonly configService: ConfigService) {}
-
-  private async _getFeedMessage(props: {
-    feedIndex: number;
-    endpoint: string;
-  }): Promise<FeedMessage> {
-    const { feedIndex, endpoint } = props;
-    const config = getConfigByFeedIndex(
-      this.configService,
-      'gtfs-realtime',
-      feedIndex,
-    );
-    const { accessKey } = config;
-    const accessKeyValue = this.configService.get(accessKey);
-    const options = {
-      method: 'GET',
-      headers: {
-        'x-api-key': accessKeyValue,
-      },
-    };
-
-    const response = await fetch(endpoint, options);
-    const buffer = await response.buffer();
-    const feedMessage = FeedMessage.decode(buffer);
-
-    return <FeedMessage>FeedMessage.toJSON(feedMessage);
-  }
+  constructor(
+    private readonly configService: ConfigService,
+    private readonly feedService: FeedService,
+  ) {}
 
   public async getAlerts(feedIndex: number): Promise<AlertEntity[]> {
     const config = getConfigByFeedIndex(
@@ -42,11 +19,19 @@ export class RealtimeService {
       feedIndex,
     );
 
+    if (!config) {
+      throw new HttpException(
+        `No valid configuration found for feedIndex: ${feedIndex}!`,
+        500,
+      );
+    }
+
     const { feedUrls } = config;
     const urls = getAlertUrls(feedUrls);
     const feeds: FeedMessage[] = await Promise.all(
       urls.map(
-        async (endpoint: string) => <FeedMessage>await this._getFeedMessage({
+        async (endpoint: string) =>
+          <FeedMessage>await this.feedService.getFeedMessage({
             feedIndex,
             endpoint,
           }),
@@ -54,6 +39,7 @@ export class RealtimeService {
     );
 
     const entities = feeds.map((feed: FeedMessage) => feed.entity);
+
     return <AlertEntity[]>(
       entities.flat().map((entity: FeedEntity) => new AlertEntity(entity.alert))
     );
